+++ src/transport/http/binary_emitter.ts
@@ -2,7 +2,8 @@ import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
 
 import { CloudEvent, Version } from "../../event/cloudevent";
 import { TransportOptions } from "../emitter";
-import { Headers, headersFor } from "./headers";
+import { Headers } from "../../messages";
+import { headersFor } from "../../messages/http/headers";
 import { asData } from "../../event/validation";
 import CONSTANTS from "../../constants";
