+++ test/integration/http_emitter_test.ts
@@ -6,7 +6,8 @@ import CONSTANTS from "../../src/constants";
 const DEFAULT_CE_CONTENT_TYPE = CONSTANTS.DEFAULT_CE_CONTENT_TYPE;
 const DEFAULT_CONTENT_TYPE = CONSTANTS.DEFAULT_CONTENT_TYPE;
 
-import { CloudEvent, Version, Emitter, Protocol, headersFor } from "../../src";
+import { CloudEvent, Version, Emitter, Protocol } from "../../src";
+import { headersFor } from "../../src/messages/http/headers";
 import { AxiosResponse } from "axios";
 
 const receiver = "https://cloudevents.io/";
